Add UsersPageView
