+++ site/src/components/UsersTable/UsersTable.tsx
@@ -1,9 +1,7 @@
-import { makeStyles } from "@material-ui/styles"
 import React from "react"
-import { Pager, UserResponse } from "../../api/types"
+import { UserResponse } from "../../api/types"
 import { Column, Table } from "../../components/Table"
 import { EmptyState } from "../EmptyState"
-import { Header } from "../Header"
 import { UserCell } from "../Table/Cells/UserCell"
 
 const Language = {
@@ -11,7 +9,6 @@ const Language = {
   usersTitle: "All users",
   emptyMessage: "No users found",
   usernameLabel: "User",
-  siteRoleLabel: "Site Role"
 }
 
 const emptyState = (
@@ -28,38 +25,19 @@ const columns: Column<UserResponse>[] = [
       return <UserCell Avatar={{ username: data.email }} primaryText={data.email} />
     },
   },
-  {
-    key: "siteRole",
-    name: Language.siteRoleLabel
-  }
 ]
 
 export interface UsersTableProps {
   users: UserResponse[]
-  pager?: Pager
 }
 
-export const UsersTable: React.FC<UsersTableProps> = ({ users, pager }) => {
-  const styles = useStyles()
+export const UsersTable: React.FC<UsersTableProps> = ({ users }) => {
   return (
-    <div className={styles.flexColumn}>
-      <Header
-        title={Language.pageTitle}
-        subTitle={`${pager?.total} total`}
-      />
-      <Table
-        columns={columns}
-        data={users}
-        title={Language.usersTitle}
-        emptyState={emptyState}
-      />
-    </div>
+    <Table
+      columns={columns}
+      data={users}
+      title={Language.usersTitle}
+      emptyState={emptyState}
+    />
   )
 }
-
-const useStyles = makeStyles(() => ({
-  flexColumn: {
-    display: "flex",
-    flexDirection: "column"
-  }
-}))
